+++ app/common/model/Type.php
@@ -15,8 +15,8 @@
 /*
 DROP TABLE IF EXISTS `nc_type`;
 CREATE TABLE IF NOT EXISTS `nc_type` (
-    `id` smallint(5) UNSIGNED NOT NULL AUTO_INCREMENT,
-    `category_id` smallint(5) UNSIGNED NOT NULL DEFAULT '0' COMMENT '栏目ID',
+    `id` int(11) UNSIGNED NOT NULL AUTO_INCREMENT,
+    `category_id` int(11) UNSIGNED NOT NULL DEFAULT '0' COMMENT '栏目ID',
     `name` varchar(100) NOT NULL DEFAULT '' COMMENT '分类名',
     `remark` varchar(300) NOT NULL DEFAULT '' COMMENT '描述',
     PRIMARY KEY (`id`),
